rename dsl to rule
