--- griffin-measure/griffin-measure-batch/src/main/scala/org/apache/griffin/measure/batch/dsl/expr/Calculatable.scala
@@ -1,10 +1,0 @@
-package org.apache.griffin.measure.batch.dsl.expr
-
-import org.apache.griffin.measure.batch.dsl.calc._
-
-
-trait Calculatable extends Serializable {
-
-  def genValue(values: Map[String, Any]): CalcValue
-
-}
