--- griffin-measure/griffin-measure-batch/src/main/scala/org/apache/griffin/measure/batch/rule/AssignRule.scala
@@ -1,7 +1,0 @@
-//package org.apache.griffin.measure.batch.rule
-//
-//import org.apache.griffin.measure.batch.dsl.expr._
-//
-//case class AssignRule(expr: AssignExpr) extends Rule {
-//
-//}
