--- griffin-measure/griffin-measure-batch/src/main/scala/org/apache/griffin/measure/batch/rule/MappingRule.scala
@@ -1,13 +1,0 @@
-//package org.apache.griffin.measure.batch.rule
-//
-//import org.apache.griffin.measure.batch.dsl.expr._
-//
-//case class MappingRule(expr: MappingExpr) extends Rule {
-//
-////  val name = ""
-////  val isKey = expr.isKey
-////
-////  val sourceRule: DataRule = _
-////  val targetRule: DataRule = _
-//
-//}
