--- griffin-measure/griffin-measure-batch/src/main/scala/org/apache/griffin/measure/batch/rule/Rule.scala
@@ -1,6 +1,0 @@
-//package org.apache.griffin.measure.batch.rule
-//
-//
-//trait Rule extends Serializable {
-//
-//}
