+++ griffin-measure/griffin-measure-batch/src/main/scala/org/apache/griffin/measure/batch/rule/calc/AnnotationValue.scala
@@ -1,4 +1,4 @@
-package org.apache.griffin.measure.batch.dsl.calc
+package org.apache.griffin.measure.batch.rule.calc
 
 // get value directly
 case class AnnotationValue(value: Option[String]) extends CalcValue {
