+++ griffin-measure/griffin-measure-batch/src/main/scala/org/apache/griffin/measure/batch/rule/calc/CalcValue.scala
@@ -1,4 +1,4 @@
-package org.apache.griffin.measure.batch.dsl.calc
+package org.apache.griffin.measure.batch.rule.calc
 
 
 trait CalcValue extends Serializable {
