+++ griffin-measure/griffin-measure-batch/src/main/scala/org/apache/griffin/measure/batch/rule/expr/Calculatable.scala
@@ -1,0 +1,10 @@
+package org.apache.griffin.measure.batch.rule.expr
+
+import org.apache.griffin.measure.batch.rule.calc._
+
+
+trait Calculatable extends Serializable {
+
+  def genValue(values: Map[String, Any]): CalcValue
+
+}
