+++ griffin-measure/griffin-measure-batch/src/main/scala/org/apache/griffin/measure/batch/rule/expr/DataExpr.scala
@@ -1,6 +1,6 @@
-package org.apache.griffin.measure.batch.dsl.expr
+package org.apache.griffin.measure.batch.rule.expr
 
-import org.apache.griffin.measure.batch.dsl.calc._
+import org.apache.griffin.measure.batch.rule.calc._
 
 trait DataExpr extends Expr with Calculatable with Recordable {
