+++ griffin-measure/griffin-measure-batch/src/main/scala/org/apache/griffin/measure/batch/rule/expr/Recordable.scala
@@ -1,4 +1,4 @@
-package org.apache.griffin.measure.batch.dsl.expr
+package org.apache.griffin.measure.batch.rule.expr
 
 
 trait Recordable extends Serializable {
