+++ griffin-measure/griffin-measure-batch/src/main/scala/org/apache/griffin/measure/batch/rule/expr/StatementAnalyzable.scala
@@ -1,4 +1,4 @@
-package org.apache.griffin.measure.batch.dsl.expr
+package org.apache.griffin.measure.batch.rule.expr
 
 
 trait StatementAnalyzable extends Serializable {
